\name{exp.norm.bertheau07}
\docType{data}
\alias{exp.norm.bertheau07}
\title{Affymetrix (U133A) Breast Cancer Expression Data from Bertheau et al 2009}
\description{
Normalized expression data of 46 samples, reduced to the 241 probesets (originally 22283 probesets),  hybridised on Affymetrix U133A chip
}
\usage{exp.norm.bertheau07}
\format{A data.frame of normalized expression data of 46 samples and 22283 probesets}
\source{ArrayExpress E-TABM-43}
\references{
Bertheau P, Turpin E, Rickman DS, Espie M, de Reynies A, Feugeas JP, et al. (2007). Exquisite
sensitivity of TP53 mutant and basal breast cancers to a dose-dense epirubicin-cyclophosphamide
regimen. PLoS Med, 4: e90.
}
\keyword{datasets}
