\name{exp.norm.chanrion08}
\docType{data}
\alias{exp.norm.chanrion08}
\title{Non Affymetrix Breast Cancer Expression dataset from Chanrion et al 2008}
\description{
Normalized expression data of 155 samples from a ER+ cohort for Tamoxifen response study, reduced to the 68 genes (originally 4397 70-mer oligonucleotides),  hybridised on Qiagen/Operon oligonucleotide microarrays
}
\usage{exp.norm.chanrion08}
\format{A data.frame of normalized expression data of 155 samples, reduced to 68 probes  with gene symbol as rownames}
\source{GEO GSE_9393}
\references{
Chanrion M, Negre V, Fontaine H, Salvetat N, Bibeau F, et al. (2008). A gene expression signature
that can predict the recurrence of tamoxifen-treated primary breast cancer. Clin Cancer Res, 14:
1744-1752.
}
\keyword{datasets}


