\name{citbcmst}
\docType{data}
\alias{citbcmst}
\title{Coreset data used to classify data in Guedj et 2011}
\description{
Breast Cancer expression dataset from CIT program projects used as a coreset to define molecular classification
}
\usage{citbcmst}
\format{A list containing the expression data of the 355 coreset samples and of the selected 375 probesets to classify data, the subtypes assigned to each sample and the Affymetrix annotation of the expression data from NetAffx (version na30)}
\source{CIT http://cit.ligue-cancer.net}
\references{
Guedj et al 2011
}
\keyword{datasets}


